Change double type aliases to num
Closes https://github.com/dart-lang/web/issues/57

Makes it easier to migrate off of dart:html APIs and allows users
to pass integers without having to convert to doubles.

+++ lib/src/helpers/events/streams.dart
@@ -368,7 +368,7 @@ class WrappedEvent {
 
   html.EventTarget? get target => wrapped.target;
 
-  double get timeStamp => wrapped.timeStamp;
+  double get timeStamp => wrapped.timeStamp.toDouble();
 
   String get type => wrapped.type;
